@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "mkdocs-publisher"
-version = "1.4.5"
+version = "2.0.0"
 description = "Publisher for MkDocs - set of plugins for content creators"
 authors = ["Maciej 'maQ' Kusz <maciej.kusz@gmail.com>"]
 license = "MIT License"
@@ -21,9 +21,7 @@ keywords = [
     "publisher",
     "tags",
 ]
-packages = [
-    {include = "mkdocs_publisher"}
-]
+packages = [{ include = "mkdocs_publisher" }]
 classifiers = [
     "Development Status :: 4 - Beta",
     "Environment :: Console",
@@ -53,19 +51,19 @@ documentation = "https://mkdocs-publisher.github.io/"
 "Bug Tracker" = "https://github.com/mkdocs-publisher/mkdocs-publisher/issues"
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.10"
 mkdocs = "^1.6.1"
-pymdown-extensions = "^10.7"
+pymdown-extensions = "^10.11.2"
 beautifulsoup4 = "^4.12.3"
-mkdocs-material = "^9.5.43"
+mkdocs-material = "^9.5.49"
 
 [tool.poetry.group.dev.dependencies]
 yamllint = "^1.35.1"
 ruff = "^0.5.7"
-pyright = "^1.1.387"
+pyright = "^1.1.391"
 
 [tool.poetry.group.test.dependencies]
-coverage = "^7.6.4"
+coverage = "^7.6.9"
 pytest = "^7.4.4"
 pytest-cov = "^4.1.0"
 pytest-check = "^2.4.1"
@@ -75,37 +73,88 @@ requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.pyright]
-extraPaths = [
-    "mkdocs_publisher",
-]
-include = [
-    "mkdocs_publisher",
-    "tests",
-]
-exclude = [
-    "docs",
-    "mkdocs_publisher_docs",
-]
-reportMissingImports = false
+extraPaths = ["mkdocs_publisher"]
+include = ["mkdocs_publisher", "tests"]
+reportMissingImports = true
 reportMissingTypeStubs = false
 reportOptionalMemberAccess = false
+reportUnusedImport = true
+reportUnusedClass = true
+reportUnusedFunction = true
+reportUnusedVariable = true
+reportUnnecessaryCast = true
+reportUnnecessaryComparison = true
+reportUnnecessaryContains = true
 typeCheckingMode = "basic"
 
 [tool.pytest.ini_options]
 addopts = "--tb=short --log-level=DEBUG --cov-report=html --cov-report=json --cov=mkdocs_publisher"
-markers = [
-    "deprecated: Tests no longer used",
-    ]
-testpaths = [
-    "tests",
-]
+markers = ["deprecated: Tests no longer used"]
+testpaths = ["tests"]
 
 [tool.ruff]
-target-version = "py39"
+target-version = "py310"
 line-length = 120
 
 [tool.ruff.lint]
-select = ["B", "E", "F", "I", "UP", "SIM"]
+select = [
+    "A",
+    # "ANN",  # TODO: uncomment
+    "ARG",
+    "B",
+    "BLE",
+    "C4",
+    "C90",
+    "DTZ",
+    "E",
+    "ERA",
+    "F",
+    "FIX",
+    "FURB",
+    "G",
+    "I",
+    "ICN",
+    "ISC",
+    "LOG",
+    "N",
+    "PERF",
+    "PGH",
+    "PL",
+    "PTH",
+    "Q",
+    "RSE",
+    "RET",
+    "RUF",
+    "S",
+    "SLF",
+    "SIM",
+    "T10",
+    "T20",
+    "TCH",
+    "TD",
+    "TID",
+    "UP",
+    "W",
+    "YTT",
+]
+ignore = [
+    "ANN101",
+    "FIX002",
+    "G004",
+    "ISC001",
+    "PLR0913",
+    "PLR2004",
+    "PLW2901",
+    "S101",
+    "TCH001",
+    "TCH002",
+    "TCH003",
+    "TD002",
+    "TD003",   # TODO: remove it
+]
+
+[tool.ruff.lint.per-file-ignores]
+"tests/*" = ["SLF"]
 
 [tool.ruff.lint.isort]
 force-single-line = true
