@@ -24,8 +24,10 @@ import logging
 import subprocess
 from hashlib import md5
 from pathlib import Path
-from typing import Optional
 from uuid import uuid4
+
+from mkdocs.config.defaults import MkDocsConfig
+from mkdocs.structure.files import Files
 
 log = logging.getLogger("mkdocs.publisher._shared.file_utils")
 
@@ -33,7 +35,7 @@ log = logging.getLogger("mkdocs.publisher._shared.file_utils")
 def run_subprocess(cmd, capture_output: bool = True) -> subprocess.CompletedProcess:
     cmd = [arg for arg in cmd if arg is not None]
     log.debug(f"Run cmd: {' '.join(cmd)}")
-    return subprocess.run(cmd, capture_output=capture_output)
+    return subprocess.run(cmd, capture_output=capture_output, check=False)  # noqa: S603
 
 
 def remove_dir(directory: Path):
@@ -45,35 +47,56 @@ def remove_dir(directory: Path):
             file.unlink(missing_ok=True)
 
 
-def calculate_file_hash(file: Path, block_size: int = 8192) -> Optional[str]:
+def calculate_file_hash(file: Path, block_size: int = 65536) -> str | None:
     try:
-        with open(file, "rb") as binary_file:
-            file_hash = md5()
+        with Path(file).open(mode="rb") as binary_file:
+            file_hash = md5()  # noqa: S324
             while chunk := binary_file.read(block_size):
                 file_hash.update(chunk)
             return file_hash.hexdigest()
-    except IsADirectoryError:
+    except (IsADirectoryError, FileNotFoundError):
         return None
+
+
+class FilesList:
+    def __init__(self, mkdocs_config: MkDocsConfig, files: Files, exclude: list[Path] | None = None):
+        self._mkdocs_config: MkDocsConfig = mkdocs_config
+        self._all_files: Files = files
+        self._exclude: list[Path] = [] if exclude is None else exclude
+
+    def list_files(self, extension: list[str], exclude: list[Path]):
+        excluded = [*exclude, *self._exclude]
+        for file in self._all_files:
+            if Path(file.dest_path).suffix.lower() in extension and not any(
+                Path(file.dest_path).is_relative_to(e) for e in excluded
+            ):
+                log.critical(file.dest_path)
 
 
 def list_files(
     directory: Path,
-    extensions: Optional[list[str]] = None,
-    exclude: Optional[list[str]] = None,
+    extensions: list[str] | None = None,
+    exclude: list[str] | None = None,
 ) -> list[Path]:
-    temp_files_list: list[Path] = []
+    files_list: list[Path] = []
     extensions = [] if extensions is None else extensions
     exclude = [] if exclude is None else exclude
+
+    excluded_files_list: list[Path] = [file.relative_to(directory) for exc in exclude for file in directory.rglob(exc)]
+
     for ext in extensions:
         for file in directory.glob(f"**/*{ext}"):
-            temp_files_list.append(file.relative_to(directory))
+            file_relative_path = file.relative_to(directory)
+            for exc in excluded_files_list:
+                log.warning(exc)
+                log.warning(str(file_relative_path).startswith(str(exc)))
 
-    excluded_files_list: list[Path] = []
-    for exc in exclude:
-        for file in directory.rglob(exc):
-            excluded_files_list.append(file.relative_to(directory))
+            files_list.extend(
+                file.relative_to(directory)
+                for exc in excluded_files_list
+                if not str(file_relative_path).startswith(str(exc))
+            )
 
-    files_list: list[Path] = [file for file in temp_files_list if file not in excluded_files_list]
     return files_list
 
 
