@@ -21,49 +21,89 @@
 # SOFTWARE.
 
 import logging
+from collections import UserDict
+from dataclasses import dataclass
+from dataclasses import field
 from pathlib import Path
-from typing import Optional
-from typing import Union
-from typing import cast
+from typing import Sequence  # noqa: UP035
 
 from mkdocs.config.defaults import MkDocsConfig
+from mkdocs.structure.nav import Link
+from mkdocs.structure.nav import Section
+from mkdocs.structure.pages import Page
 
 from mkdocs_publisher._shared import mkdocs_utils
-from mkdocs_publisher.blog.config import BlogPluginConfig
+from mkdocs_publisher.meta.config import MetaPluginConfig
 from mkdocs_publisher.obsidian.config import ObsidianPluginConfig
 
 log = logging.getLogger("mkdocs.publisher._shared.publisher_utils")
 
 
-def get_blog_dir(mkdocs_config: MkDocsConfig) -> Optional[Path]:
-    blog_config: Optional[Union[BlogPluginConfig, Path]] = cast(
-        BlogPluginConfig,
-        mkdocs_utils.get_plugin_config(
-            mkdocs_config=mkdocs_config,
-            plugin_name="pub-blog",
-        ),
-    )
-    if blog_config is not None:
-        blog_config = Path(mkdocs_config.docs_dir).joinpath(blog_config.blog_dir)
-
-    return blog_config
-
-
-def get_obsidian_dirs(mkdocs_config: MkDocsConfig) -> tuple[list[Path], Optional[Path]]:
+def get_obsidian_dirs(mkdocs_config: MkDocsConfig) -> tuple[list[Path], Path | None]:
     ignored_dirs: list[Path] = []
-    attachments_dir: Optional[Path] = None
+    attachments_dir: Path | None = None
     docs_dir = Path(mkdocs_config.docs_dir)
 
-    obsidian_config: Optional[ObsidianPluginConfig] = cast(
-        ObsidianPluginConfig,
-        mkdocs_utils.get_plugin_config(
-            mkdocs_config=mkdocs_config,
-            plugin_name="pub-obsidian",
-        ),
-    )
-    if obsidian_config is not None:
+    obsidian_config: ObsidianPluginConfig | None = mkdocs_utils.get_plugin_config(
+        plugin_config_type=ObsidianPluginConfig,  # type: ignore[reportArgumentType]
+        mkdocs_config=mkdocs_config,
+    )  # type: ignore[reportAssignmentType]
+
+    if obsidian_config:
         ignored_dirs.append(docs_dir.joinpath(obsidian_config.obsidian_dir))
         ignored_dirs.append(docs_dir.joinpath(obsidian_config.templates_dir))
         attachments_dir = docs_dir.joinpath(obsidian_config.attachments_dir)
 
     return ignored_dirs, attachments_dir
+
+
+@dataclass
+class PublisherFile:
+    path: Path
+    abs_path: Path = field(repr=False)
+    is_dir: bool = field(default=False)
+    is_draft: bool | None = field(default=None)
+    slug: str | None = field(default=None)
+    title: str | None = field(default=None)
+
+
+class PublisherFiles(UserDict):
+    def __init__(self):
+        self._on_serve: bool = False
+        self._mkdocs_config: MkDocsConfig | None = None
+        self._meta_plugin_config: MetaPluginConfig | None = None
+
+        super().__init__()
+
+    @property
+    def on_serve(self) -> bool:
+        return self._on_serve
+
+    @on_serve.setter
+    def on_serve(self, on_serve: bool):
+        self._on_serve = on_serve
+
+    def set_configs(self, mkdocs_config: MkDocsConfig, meta_plugin_config: MetaPluginConfig):
+        self._mkdocs_config = mkdocs_config
+        self._meta_plugin_config = meta_plugin_config
+
+    def add_files(self):
+        raise NotImplementedError  # pragma: no cover
+
+
+def nav_cleanup(items, removal_list: Sequence[str | Path]) -> list:
+    removal_list = [str(p) for p in removal_list]
+    nav = []
+    for item in items:
+        if isinstance(item, Section):
+            item.children = nav_cleanup(items=item.children, removal_list=removal_list)
+            # If section is empty, skip it
+            if len(item.children) > 0:
+                nav.append(item)
+        elif (
+            isinstance(item, Page)
+            and str(item.file.src_path) not in removal_list
+            and str(Path(item.file.src_path).parent) not in removal_list
+        ) or (isinstance(item, Link) and item.title not in removal_list):
+            nav.append(item)
+    return nav
