@@ -27,7 +27,6 @@ from collections import OrderedDict
 from pathlib import Path
 from typing import Any
 from typing import Literal
-from typing import Optional
 from typing import cast
 
 from mkdocs.config import Config
@@ -44,6 +43,7 @@ from mkdocs_publisher._shared import resources
 from mkdocs_publisher.blog import creators
 from mkdocs_publisher.blog import modifiers
 from mkdocs_publisher.blog import parsers
+from mkdocs_publisher.blog.blog_files import BlogFiles
 from mkdocs_publisher.blog.config import BlogPluginConfig
 from mkdocs_publisher.blog.structures import BlogConfig
 from mkdocs_publisher.obsidian.md_links import MarkdownLinks
@@ -52,16 +52,28 @@ log = logging.getLogger("mkdocs.publisher.blog.plugin")
 
 
 class BlogPlugin(BasePlugin[BlogPluginConfig]):
-    def __init__(self):
-        self.blog_config = BlogConfig()  # Empty instance
-        self._start_page: bool = False
-        self._on_serve: bool = False
+    supports_multiple_instances = False  # TODO: add multiple instances support (require changes in meta plugin)
 
-    def on_startup(self, *, command: Literal["build", "gh-deploy", "serve"], dirty: bool) -> None:
+    def __init__(self):
+        self._on_serve: bool = False
+        self._start_page: bool = False
+        self._blog_files: BlogFiles = BlogFiles()
+
+        # ==== Old below
+
+        self.blog_config = BlogConfig()  # Empty instance
+
+    def on_startup(self, *, command: Literal["build", "gh-deploy", "serve"], dirty: bool) -> None:  # noqa: ARG002
         if command == "serve":
             self._on_serve = True
+        self._blog_files.on_serve = self._on_serve
 
     def on_config(self, config: MkDocsConfig) -> Config:
+        self._blog_files.set_configs(mkdocs_config=config, blog_plugin_config=self.config)
+        self._blog_files.add_blog_files()
+
+        # ==== Old below
+
         # Initialization of all the values
         self.blog_config.parse_configs(mkdocs_config=config, plugin_config=self.config)
 
@@ -71,7 +83,7 @@ class BlogPlugin(BasePlugin[BlogPluginConfig]):
 
         # Detect if blog is a starting page
         if len(config.nav) > 0:
-            first_value = list(config.nav[0].values())[0]
+            first_value = next(iter(config.nav[0].values()))
             if isinstance(first_value, str) and first_value == str(self.blog_config.blog_dir):
                 self._start_page = True
 
@@ -109,7 +121,7 @@ class BlogPlugin(BasePlugin[BlogPluginConfig]):
         config.nav = new_nav
         return config
 
-    def on_nav(self, nav: Navigation, config: MkDocsConfig, files: Files) -> Navigation:
+    def on_nav(self, nav: Navigation, config: MkDocsConfig, files: Files) -> Navigation:  # noqa: ARG002
         modifiers.blog_post_nav_remove(start_page=self._start_page, blog_config=self.blog_config, nav=nav)
 
         return nav
@@ -123,8 +135,13 @@ class BlogPlugin(BasePlugin[BlogPluginConfig]):
 
     @event_priority(-100)  # Run after all other plugins
     def on_page_context(
-        self, context: dict[str, Any], *, page: Page, config: MkDocsConfig, nav: Navigation
-    ) -> Optional[dict[str, Any]]:
+        self,
+        context: dict[str, Any],
+        *,
+        page: Page,
+        config: MkDocsConfig,  # noqa: ARG002
+        nav: Navigation,  # noqa: ARG002
+    ) -> dict[str, Any] | None:
         if Path(page.file.src_path).parts[0] == self.config.blog_dir:
             page.meta[self.config.comments.key_name] = self.config.comments.enabled
 
@@ -136,7 +153,7 @@ class BlogPlugin(BasePlugin[BlogPluginConfig]):
         return context
 
     @event_priority(-100)  # Run after all other plugins
-    def on_page_markdown(self, markdown: str, *, page: Page, config: MkDocsConfig, files: Files) -> Optional[str]:
+    def on_page_markdown(self, markdown: str, *, page: Page, config: MkDocsConfig, files: Files) -> str | None:  # noqa: ARG002
         md_links = MarkdownLinks(mkdocs_config=config)
         markdown = md_links.normalize_relative_links(
             markdown=markdown,
@@ -162,11 +179,25 @@ class BlogPlugin(BasePlugin[BlogPluginConfig]):
         return markdown
 
     @event_priority(-100)  # Run after all other plugins
-    def on_build_error(self, error: Exception) -> None:
+    def on_post_build(self, *, config: MkDocsConfig) -> None:  # noqa: ARG002
+        self._blog_files.remove_temp_dirs()
+
+        # ==== Old below
+        with contextlib.suppress(AttributeError):
+            file_utils.remove_dir(directory=self.blog_config.temp_dir)
+
+    @event_priority(-100)  # Run after all other plugins
+    def on_build_error(self, error: Exception) -> None:  # noqa: ARG002
+        self._blog_files.remove_temp_dirs()
+
+        # ==== Old below
         with contextlib.suppress(AttributeError):
             file_utils.remove_dir(directory=self.blog_config.temp_dir)
 
     @event_priority(-100)  # Run after all other plugins
     def on_shutdown(self) -> None:
+        self._blog_files.remove_temp_dirs()
+
+        # ==== Old below
         with contextlib.suppress(AttributeError):
             file_utils.remove_dir(directory=self.blog_config.temp_dir)
