@@ -25,7 +25,6 @@ from dataclasses import dataclass
 from dataclasses import field
 from datetime import datetime
 from pathlib import Path
-from typing import Optional
 
 from mkdocs.config.defaults import MkDocsConfig
 
@@ -40,11 +39,11 @@ class BlogPost:
 
     title: str
     date: datetime
-    path: Optional[str]
-    content: Optional[str]
-    tags: Optional[list[str]]
-    categories: Optional[list[str]]
-    slug: Optional[str] = None
+    path: str | None
+    content: str | None
+    tags: list[str] | None
+    categories: list[str] | None
+    slug: str | None = None
     teaser: str = ""
     is_teaser: bool = False
 
@@ -77,20 +76,20 @@ class Translation:
 class BlogConfig:
     mkdocs_config: MkDocsConfig = field(init=False)
     plugin_config: BlogPluginConfig = field(init=False)
-    meta_config: Optional[MetaPluginConfig] = field(init=False, default=None)
+    meta_config: MetaPluginConfig | None = field(init=False, default=None)
     translation: Translation = field(init=False)
     temp_dir: Path = field(init=False)
     docs_dir: Path = field(init=False)
     blog_dir: Path = field(init=False)
     site_dir: Path = field(init=False)
-    blog_posts: dict[datetime, BlogPost] = field(init=False, default_factory=lambda: dict())
-    temp_files: dict[str, Path] = field(init=False, default_factory=lambda: dict())
+    blog_posts: dict[datetime, BlogPost] = field(init=False, default_factory=lambda: {})
+    temp_files: dict[str, Path] = field(init=False, default_factory=lambda: {})
 
     @property
     def temp_files_list(self) -> list[str]:
         temp_files = []
         for path in self.temp_files.values():
-            temp_files.append(str(path.relative_to(self.temp_dir)))
+            temp_files.append(str(path.relative_to(self.temp_dir)))  # noqa: PERF401
         return temp_files
 
     def parse_configs(self, mkdocs_config: MkDocsConfig, plugin_config: BlogPluginConfig):
@@ -98,9 +97,10 @@ class BlogConfig:
 
         self.mkdocs_config = mkdocs_config
         self.plugin_config = plugin_config
-        self.meta_config: Optional[MetaPluginConfig] = (
-            mkdocs_utils.get_plugin_config(mkdocs_config=mkdocs_config, plugin_name="pub-meta") or None
-        )  # type: ignore
+        self.meta_config: MetaPluginConfig | None = mkdocs_utils.get_plugin_config(
+            plugin_config_type=MetaPluginConfig,  # type: ignore[reportArgumentType]
+            mkdocs_config=mkdocs_config,
+        )  # type: ignore[reportAttributeAccessIssue]
         self.temp_dir = Path(plugin_config.temp_dir)
         self.docs_dir = Path(mkdocs_config.docs_dir)
         self.blog_dir = Path(plugin_config.blog_dir)
