@@ -22,99 +22,69 @@
 
 import logging
 import re
-from collections import UserDict
 from collections.abc import Generator
 from dataclasses import dataclass
 from dataclasses import field
 from functools import cached_property
 from pathlib import Path
 from typing import Any
-from typing import Optional
 from urllib.parse import quote
 
-from mkdocs.config.defaults import MkDocsConfig
 from mkdocs.structure.files import File
 from mkdocs.structure.files import Files
-from mkdocs.utils import meta as meta_parser
 
 from mkdocs_publisher._shared import links
+from mkdocs_publisher._shared import mkdocs_utils
+from mkdocs_publisher._shared import publisher_utils
 from mkdocs_publisher._shared import templates
-from mkdocs_publisher._shared.urls import create_slug
-from mkdocs_publisher.meta.config import MetaPluginConfig
 from mkdocs_publisher.meta.config import OverviewChoiceEnum
 from mkdocs_publisher.meta.config import PublishChoiceEnum
 from mkdocs_publisher.meta.config import TitleChoiceEnum
 
-log = logging.getLogger("mkdocs.publisher._shared.meta_files")
+log = logging.getLogger("mkdocs.publisher.meta.meta_files")
 
 
 HEADINGS_RE = re.compile(r"^#+ (?P<title>[^|#\r\n\t\f\v]+)$")
 
 
 @dataclass
-class MetaFile:
-    is_dir: bool
-    path: Path
-    abs_path: Path = field(repr=False)
-    file: Optional[File] = field(default=None, repr=False)
-    is_draft: Optional[bool] = field(default=None)
+class MetaFile(publisher_utils.PublisherFile):
+    file: File | None = field(default=None, repr=False)
     is_hidden: bool = field(default=False)
     is_overview: bool = field(default=False)
-    redirect: Optional[str] = field(default=None)
-    slug: Optional[str] = field(default=None)
-    title: Optional[str] = field(default=None)
-    url: Optional[str] = field(default=None)
+    redirect: str | None = field(default=None)
+    url: str | None = field(default=None)
 
     @property
     def name(self) -> str:
         return self.path.name
 
     @property
-    def parent(self) -> Optional[Path]:
+    def parent(self) -> Path | None:
         parent = self.path.parent
         if str(parent) == ".":
             parent = None
         return parent
 
 
-class MetaFiles(UserDict):
+class MetaFiles(publisher_utils.PublisherFiles):
     def __init__(self):
-        self._on_serve: bool = False
-        self._mkdocs_config: Optional[MkDocsConfig] = None
-        self._meta_plugin_config: Optional[MetaPluginConfig] = None
         self._hidden_paths: list[Path] = []
+
         super().__init__()
 
     @property
     def dir_meta_file(self) -> str:
         return self._meta_plugin_config.dir_meta_file
 
-    @property
-    def on_serve(self) -> bool:
-        return self._on_serve
-
-    @on_serve.setter
-    def on_serve(self, on_serve: bool):
-        self._on_serve = on_serve
-
-    def add_hidden_path(self, hidden_path: Optional[Path]):
+    def add_hidden_path(self, hidden_path: Path | None):
         if hidden_path is not None:
             self._hidden_paths.append(hidden_path.relative_to(self._mkdocs_config.docs_dir))
-
-    def set_configs(self, mkdocs_config: MkDocsConfig, meta_plugin_config: MetaPluginConfig):
-        self._mkdocs_config = mkdocs_config
-        self._meta_plugin_config = meta_plugin_config
-
-    @staticmethod
-    def _read_md_file(meta_file_path: Path) -> tuple[str, dict[str, Any]]:  # pragma: no cover
-        with meta_file_path.open(encoding="utf-8-sig", errors="strict") as md_file:
-            # Add empty line at the end of file and read all data
-            return meta_parser.get_data(f"{md_file.read()}\n")
 
     def _get_title(self, meta_file: MetaFile, meta: dict[str, Any], markdown: str):
         """Calculate title for given file"""
 
-        title: Optional[str] = None
+        title: str | None = None
         mode = self._meta_plugin_config.title.mode
         if mode == TitleChoiceEnum.META:
             title = meta.get(self._meta_plugin_config.title.key_name)
@@ -125,10 +95,10 @@ class MetaFiles(UserDict):
             ):
                 log.warning(
                     f'Title value from "{self._meta_plugin_config.title.key_name}" meta data '
-                    f'is missing for file: "{str(meta_file.path)}"'
+                    f'is missing for file: "{meta_file.path!s}"',
                 )
 
-        if title is None and (mode == TitleChoiceEnum.META or mode == TitleChoiceEnum.HEAD):
+        if title is None and mode in {TitleChoiceEnum.META.name, TitleChoiceEnum.HEAD.name}:
             headings = re.findall(HEADINGS_RE, markdown)
             title = str(headings[0]).strip() if len(headings) > 0 else None
             if (
@@ -136,9 +106,9 @@ class MetaFiles(UserDict):
                 and self._meta_plugin_config.title.warn_on_missing_header
                 and not (meta_file.is_draft or (meta_file.is_dir and not meta_file.is_overview))
             ):
-                log.warning(f'Title value from first heading is missing for file: "{str(meta_file.path)}"')
+                log.warning(f'Title value from first heading is missing for file: "{meta_file.path!s}"')
 
-        if title is None and (mode == TitleChoiceEnum.META or mode == TitleChoiceEnum.FILE):
+        if title is None and mode in {TitleChoiceEnum.META.name, TitleChoiceEnum.FILE.name}:
             title = str(meta_file.path.stem).replace("_", " ").title()
 
         meta_file.title = title
@@ -146,7 +116,7 @@ class MetaFiles(UserDict):
     def _get_slug(self, meta_file: MetaFile, meta: dict[str, Any]):
         """Calculate slug for given file"""
 
-        meta_file.slug = create_slug(  # pragma: no cover
+        meta_file.slug = links.create_slug(  # pragma: no cover
             file_name=meta_file.path.stem,
             slug_mode=self._meta_plugin_config.slug.mode,
             slug=meta.get(self._meta_plugin_config.slug.key_name),
@@ -183,12 +153,12 @@ class MetaFiles(UserDict):
             # Document with redirection should be hidden
             meta[self._meta_plugin_config.publish.key_name] = PublishChoiceEnum.HIDDEN.name.lower()
 
-        meta_file.redirect = redirect  # type: ignore
+        meta_file.redirect = redirect
 
         return meta
 
     def _get_overview(self, meta_file: MetaFile, meta: dict[str, Any], markdown: str):
-        """Overview works only for metafiles ("README.md", "index.md") that are detected as dir"""
+        """Overview works only for meta files ("README.md", "index.md") that are detected as dir"""
 
         if not meta_file.is_dir:
             meta_file.is_overview = False
@@ -202,28 +172,27 @@ class MetaFiles(UserDict):
             else:
                 meta_file.is_overview = is_overview
 
-    def _get_publish_status(self, meta_file: MetaFile, meta: dict[str, Any]):
+    def _get_publish_status(self, meta_file: MetaFile, meta: dict[str, Any]):  # noqa: C901
         """Calculate publication status for given file"""
 
         publish = meta.get(str(self._meta_plugin_config.publish.key_name), None)
         # Get default values if publish status is not specified
-        if publish is None:
-            if meta_file.is_dir:
-                publish = self._meta_plugin_config.publish.dir_default
-                if self._meta_plugin_config.publish.dir_warn_on_missing:
-                    log.warning(
-                        f'Missing "{self._meta_plugin_config.publish.key_name}" value in '
-                        f'file "{meta_file.path}". Setting to '
-                        f'default value: "{self._meta_plugin_config.publish.dir_default}".'
-                    )
-            else:
-                publish = self._meta_plugin_config.publish.file_default
-                if self._meta_plugin_config.publish.file_warn_on_missing:
-                    log.warning(
-                        f'Missing "{self._meta_plugin_config.publish.key_name}" value in '
-                        f'file "{meta_file.path}". Setting to '
-                        f'default value: "{self._meta_plugin_config.publish.file_default}".'
-                    )
+        if publish is None and meta_file.is_dir:
+            publish = self._meta_plugin_config.publish.dir_default
+            if self._meta_plugin_config.publish.dir_warn_on_missing:
+                log.warning(
+                    f'Missing "{self._meta_plugin_config.publish.key_name}" value in '
+                    f'file "{meta_file.path}". Setting to '
+                    f'default value: "{self._meta_plugin_config.publish.dir_default}".',
+                )
+        elif publish is None:
+            publish = self._meta_plugin_config.publish.file_default
+            if self._meta_plugin_config.publish.file_warn_on_missing:
+                log.warning(
+                    f'Missing "{self._meta_plugin_config.publish.key_name}" value in '
+                    f'file "{meta_file.path}". Setting to '
+                    f'default value: "{self._meta_plugin_config.publish.file_default}".',
+                )
 
         # Override some values inherited from parent
         if meta_file.parent is not None:
@@ -231,7 +200,7 @@ class MetaFiles(UserDict):
             if meta_file_parent.is_draft:
                 publish = False
             if meta_file_parent.is_hidden and publish not in PublishChoiceEnum.drafts():
-                publish = PublishChoiceEnum.HIDDEN.name.lower()
+                publish = PublishChoiceEnum.HIDDEN.name
 
         # When live preview is running, all pages are visible
         if self._on_serve and not meta_file.is_hidden:
@@ -241,11 +210,11 @@ class MetaFiles(UserDict):
             publish = self._meta_plugin_config.publish.file_default
             log.warning(
                 f'Wrong key "{self._meta_plugin_config.publish.key_name}" value ({publish}) '
-                f'in file "{meta_file.path}" (only {PublishChoiceEnum.choices()} are possible)'
+                f'in file "{meta_file.path}" (only {PublishChoiceEnum.choices()} are possible)',
             )
 
         # Set values depends on publish status
-        if meta_file.path in self._hidden_paths or publish == PublishChoiceEnum.HIDDEN:
+        if meta_file.path in self._hidden_paths or publish == PublishChoiceEnum.HIDDEN.name:
             meta_file.is_hidden = True
             meta_file.is_draft = False
         elif publish in PublishChoiceEnum.published():
@@ -258,7 +227,7 @@ class MetaFiles(UserDict):
     def _get_metadata(self, meta_file: MetaFile, meta_file_path: Path):  # pragma: no cover
         """Read all metadata values for given file"""
 
-        markdown, meta = self._read_md_file(meta_file_path=meta_file_path)
+        markdown, meta = mkdocs_utils.read_md_file(md_file_path=meta_file_path)
 
         # Order of method execution is crucial for reading all values.
         meta = self._get_redirect(meta_file=meta_file, meta=meta, markdown=markdown)
@@ -272,17 +241,13 @@ class MetaFiles(UserDict):
         """Add file"""
 
         if meta_file.is_dir:
-            meta_file_exists = False
-
             # Calculate properties based on meta file metadata
             meta_file_path = meta_file.abs_path.joinpath(self._meta_plugin_config.dir_meta_file)
             if meta_file_path.exists():
-                meta_file_exists = True
                 self._get_metadata(meta_file=meta_file, meta_file_path=meta_file_path)
-
-            if not meta_file_exists:
+            else:
                 meta_file.title = str(meta_file.path.stem)
-                meta_file.slug = create_slug(
+                meta_file.slug = links.create_slug(
                     file_name=str(meta_file.name),
                     slug_mode=self._meta_plugin_config.slug.mode,
                     slug=meta_file.path.stem,
@@ -292,6 +257,7 @@ class MetaFiles(UserDict):
                 meta_file.is_draft = not self._meta_plugin_config.publish.dir_default
         else:
             self._get_metadata(meta_file=meta_file, meta_file_path=meta_file.abs_path)
+
         super().__setitem__(path, meta_file)
 
     def _drafts(self, files: bool = False, dirs: bool = False) -> dict[str, MetaFile]:
@@ -348,15 +314,17 @@ class MetaFiles(UserDict):
     def hidden(self) -> dict[str, MetaFile]:
         return self._hidden()
 
-    def add_meta_files(self, ignored_dirs: list[Path]):
+    def add_files(self, ignored_dirs: list[Path] | None = None):
         """Iterate over all files and directories in docs directory"""
+        if not ignored_dirs:
+            ignored_dirs = []
 
         for docs_file in sorted(Path(self._mkdocs_config.docs_dir).rglob("*")):
-            meta_link: Optional[MetaFile] = None
-            is_ignored = any([docs_file.is_relative_to(ignored_dir) for ignored_dir in ignored_dirs])
+            meta_file: MetaFile | None = None
+            is_ignored = any(docs_file.is_relative_to(ignored_dir) for ignored_dir in ignored_dirs)
 
             if not is_ignored and docs_file.is_dir():
-                meta_link = MetaFile(
+                meta_file = MetaFile(
                     path=docs_file.relative_to(self._mkdocs_config.docs_dir),
                     abs_path=docs_file,
                     is_dir=True,
@@ -366,14 +334,14 @@ class MetaFiles(UserDict):
                 and docs_file.suffix == ".md"
                 and docs_file.name != self._meta_plugin_config.dir_meta_file
             ):
-                meta_link = MetaFile(
+                meta_file = MetaFile(
                     path=docs_file.relative_to(self._mkdocs_config.docs_dir),
                     abs_path=docs_file,
                     is_dir=False,
                 )
 
-            if meta_link is not None:
-                self[str(meta_link.path)] = meta_link
+            if meta_file is not None:
+                self[str(meta_file.path)] = meta_file
 
     def _change_file_slug(self, file: File, file_path: Path):
         # Get URL parts
@@ -391,7 +359,7 @@ class MetaFiles(UserDict):
 
         # Replace URL parts that have slug defined
         for position, path_part in enumerate(path_parts):
-            meta_file: Optional[MetaFile] = self.get(str(path_part), None)
+            meta_file: MetaFile | None = self.get(str(path_part), None)
             if meta_file is not None:
                 url_parts[position] = str(meta_file.slug)
 
@@ -416,7 +384,7 @@ class MetaFiles(UserDict):
             file_path: Path = Path(file.src_path)
             if (
                 (
-                    not any([Path(str(file.abs_src_path)).is_relative_to(d) for d in ignored_dirs])
+                    not any(Path(str(file.abs_src_path)).is_relative_to(d) for d in ignored_dirs)
                     and file.src_path not in self.draft_files
                     and str(file_path.name) != self._meta_plugin_config.dir_meta_file
                 )
@@ -434,7 +402,7 @@ class MetaFiles(UserDict):
                 self[file.src_path].url = file.url
         return new_files
 
-    def generate_redirect_page(self, file: File) -> Optional[str]:
+    def generate_redirect_page(self, file: File) -> str | None:
         """Generates content of redirect page"""
 
         meta_file: MetaFile = self[file.src_path]
@@ -445,8 +413,7 @@ class MetaFiles(UserDict):
                 "url": f"{self._mkdocs_config.site_url}{self[meta_file.redirect].url}",
             }
             return templates.render(tpl_file="redirect.html", context=redirect_context)
-        else:
-            return None
+        return None
 
     def clean_redirect_files(self, files: Files) -> Files:
         """Remove documents that are just redirects to an external URL"""
