@@ -25,7 +25,6 @@ import logging
 from datetime import datetime
 from pathlib import Path
 from typing import Literal
-from typing import Optional
 
 from mkdocs.config import Config
 from mkdocs.config.defaults import MkDocsConfig
@@ -35,7 +34,9 @@ from mkdocs.structure.files import Files
 from mkdocs.structure.nav import Navigation
 from mkdocs.structure.pages import Page
 
+from mkdocs_publisher._shared import mkdocs_utils
 from mkdocs_publisher._shared import publisher_utils
+from mkdocs_publisher.blog.config import BlogPluginConfig
 from mkdocs_publisher.meta.config import MetaPluginConfig
 from mkdocs_publisher.meta.meta_files import MetaFiles
 from mkdocs_publisher.meta.meta_nav import MetaNav
@@ -44,66 +45,74 @@ log = logging.getLogger("mkdocs.publisher.meta.plugin")
 
 
 class MetaPlugin(BasePlugin[MetaPluginConfig]):
+    supports_multiple_instances = False
+
     def __init__(self):
-        self._on_serve = False
-        self._attachments_dir: Optional[Path] = None
+        self._on_serve: bool = False
+        self._attachments_dir: Path | None = None
         self._ignored_dirs: list[Path] = []
         self._meta_files: MetaFiles = MetaFiles()
-        self._meta_nav: Optional[MetaNav] = None
+        self._meta_nav: MetaNav | None = None
 
     def on_startup(self, *, command: Literal["build", "gh-deploy", "serve"], dirty: bool) -> None:  # pragma: no cover
+        _ = dirty
         if command == "serve":
             self._on_serve = True
         self._meta_files.on_serve = self._on_serve
 
     @event_priority(100)  # Run before any other plugins
-    def on_config(self, config: MkDocsConfig) -> Optional[Config]:  # pragma: no cover
+    def on_config(self, config: MkDocsConfig) -> Config | None:  # pragma: no cover
         # Set some default values
         log.info("Read files and directories metadata")
-        blog_dir: Optional[Path] = publisher_utils.get_blog_dir(mkdocs_config=config)
+        blog_config: BlogPluginConfig = mkdocs_utils.get_plugin_config(
+            plugin_config_type=BlogPluginConfig,  # type: ignore[reportArgumentType]
+            mkdocs_config=config,
+        )  # type: ignore[reportAssignmentType]
         self._meta_nav = MetaNav(
             meta_files=self._meta_files,
-            blog_dir=blog_dir.relative_to(config.docs_dir) if blog_dir else blog_dir,
+            blog_dir=blog_config.blog_dir if blog_config is not None else None,  # type: ignore[reportArgumentType]
         )
         self._ignored_dirs, self._attachments_dir = publisher_utils.get_obsidian_dirs(mkdocs_config=config)
         self._meta_files.set_configs(mkdocs_config=config, meta_plugin_config=self.config)
         self._meta_files.add_hidden_path(hidden_path=self._attachments_dir)
-        self._meta_files.add_meta_files(ignored_dirs=self._ignored_dirs)
+        self._meta_files.add_files(ignored_dirs=self._ignored_dirs)
 
-        log.info(f"Ignored directories: " f"{[str(d.relative_to(config.docs_dir)) for d in self._ignored_dirs]}")
-        log.info(f"Draft files and directories: " f"{list(self._meta_files.drafts.keys())}")
-        log.info(f"Hidden files and directories: " f"{list(self._meta_files.hidden.keys())}")
+        log.info(f"Ignored directories: {[str(d.relative_to(config.docs_dir)) for d in self._ignored_dirs]}")
+        log.info(f"Draft files and directories: {list(self._meta_files.drafts.keys())}")
+        log.info(f"Hidden files and directories: {list(self._meta_files.hidden.keys())}")
 
         config.nav = self._meta_nav.build_nav(mkdocs_config=config)
 
         return config
 
     @event_priority(-100)
-    def on_files(self, files: Files, *, config: MkDocsConfig) -> Optional[Files]:  # pragma: no cover
+    def on_files(self, files: Files, *, config: MkDocsConfig) -> Files | None:  # pragma: no cover
+        _ = config
         new_files = self._meta_files.clean_redirect_files(files=files)
         new_files = self._meta_files.change_files_slug(files=new_files, ignored_dirs=self._ignored_dirs)
-        new_files = self._meta_files.clean_draft_files(files=new_files)
+        return self._meta_files.clean_draft_files(files=new_files)
 
-        return new_files
-
-    def on_nav(
-        self, nav: Navigation, *, config: MkDocsConfig, files: Files
-    ) -> Optional[Navigation]:  # pragma: no cover
-        removal_list = [*self._meta_files.drafts.keys(), *self._meta_files.hidden.keys()]
+    @event_priority(-99)
+    def on_nav(self, nav: Navigation, *, config: MkDocsConfig, files: Files) -> Navigation | None:  # pragma: no cover
+        _, _ = config, files
+        removal_list = [
+            *self._meta_files.drafts.keys(),
+            *self._meta_files.hidden.keys(),
+        ]
         log.debug(f"Nav elements to remove: {removal_list}")
-        nav.items = self._meta_nav.nav_cleanup(
-            items=nav.items,
-            removal_list=removal_list,
-        )
+        # TODO: create prev/next page cleanup for hidden pages
+
+        nav.items = publisher_utils.nav_cleanup(items=nav.items, removal_list=removal_list)
         return nav
 
     @event_priority(-100)  # Run after all other plugins
-    def on_page_markdown(self, markdown: str, *, page: Page, config: MkDocsConfig, files: Files):
+    def on_page_markdown(self, markdown: str, *, page: Page, config: MkDocsConfig, files: Files):  # noqa: ARG002
         # Modify page update date
         # TODO: move date format to config
         # TODO: warn on missing in config
         update_date: datetime = page.meta.get(
-            "update", page.meta.get("date", datetime.strptime(page.update_date, "%Y-%m-%d"))
+            "update",
+            page.meta.get("date", datetime.strptime(page.update_date, "%Y-%m-%d")),  # noqa: DTZ007
         )
         with contextlib.suppress(AttributeError):
             page.update_date = update_date.strftime("%Y-%m-%d")
@@ -115,9 +124,16 @@ class MetaPlugin(BasePlugin[MetaPluginConfig]):
             page.meta["search"] = {"exclude": True}
 
     @event_priority(-100)  # Run after all other plugins
-    def on_post_page(self, output: str, *, page: Page, config: MkDocsConfig) -> Optional[str]:  # pragma: no cover
+    def on_post_page(self, output: str, *, page: Page, config: MkDocsConfig) -> str | None:  # pragma: no cover
+        _ = config
         if page.file.src_path in self._meta_files:
-            redirect_page: Optional[str] = self._meta_files.generate_redirect_page(file=page.file)
+            redirect_page: str | None = self._meta_files.generate_redirect_page(file=page.file)
             if redirect_page:
                 output = redirect_page
         return output
+
+    # TODO: sitemap modification for hidden, draft and by frontmatter
+    def on_post_template(self, output_content: str, /, *, template_name: str, config: MkDocsConfig) -> str | None:  # noqa: ARG002
+        if template_name == "sitemap.xml":
+            log.debug(template_name)
+            # log.debug(output_content)  # noqa: ERA001
