@@ -22,21 +22,24 @@
 
 import logging
 from pathlib import Path
-from typing import Optional
+from typing import TYPE_CHECKING
 
 from mkdocs_publisher._shared import file_utils
-from mkdocs_publisher.minifier import config as minifier_config
 from mkdocs_publisher.minifier.base import BaseMinifier
 from mkdocs_publisher.minifier.base import CachedFile
+
+if TYPE_CHECKING:  # pragma: no cover
+    from mkdocs_publisher.minifier import config as minifier_config
 
 log = logging.getLogger("mkdocs.publisher.minifier.minifiers")
 
 
 def _is_cmd_installed(cmd: list[str]) -> bool:
     try:
-        result = file_utils.run_subprocess(cmd=cmd)
+        cmd_output = file_utils.run_subprocess(cmd=cmd)
+        log.debug(cmd_output.stdout.decode("utf-8") if cmd_output.stdout else cmd_output.stderr.decode("utf-8"))
 
-        if result.returncode != 0:
+        if cmd_output.returncode != 0:
             return False
     except FileNotFoundError:
         return False
@@ -62,7 +65,7 @@ class PngMinifier(BaseMinifier):
             are_installed = False
         return are_installed
 
-    def minifier(self, cached_file: CachedFile) -> Optional[CachedFile]:
+    def minifier(self, cached_file: CachedFile) -> CachedFile | None:
         try:
             input_file = self._mkdocs_config.site_dir / cached_file.original_file_path
             output_file = self._plugin_config.cache_dir / cached_file.cached_file_name
@@ -125,23 +128,23 @@ class JpegMinifier(BaseMinifier):
             log.warning("Mozilla JPEG Encoder is not installed.")
         return are_installed
 
-    def minifier(self, cached_file: CachedFile) -> Optional[CachedFile]:
+    def minifier(self, cached_file: CachedFile) -> CachedFile | None:
         try:
             input_file = self._mkdocs_config.site_dir / cached_file.original_file_path
             output_file = self._plugin_config.cache_dir / cached_file.cached_file_name
 
-            djpg_cmd = [
+            djpeg_cmd = [
                 self._minify_options.djpeg_path,
                 "-targa",
                 "-outfile",
                 str(output_file.with_suffix(".tga")),
                 str(input_file),
             ]
-            if file_utils.run_subprocess(cmd=djpg_cmd).returncode != 0:
+            if file_utils.run_subprocess(cmd=djpeg_cmd).returncode != 0:
                 output_file.with_suffix(".tga").unlink(missing_ok=True)
                 return None
 
-            cjpg_cmd = [
+            cjpeg_cmd = [
                 self._minify_options.cjpeg_path,
                 "-targa",
                 "-smooth" if int(self._minify_options.smooth) > 0 else None,
@@ -152,7 +155,7 @@ class JpegMinifier(BaseMinifier):
                 str(output_file.with_suffix(".jpg_")),
                 str(output_file.with_suffix(".tga")),
             ]
-            if file_utils.run_subprocess(cmd=cjpg_cmd).returncode != 0:
+            if file_utils.run_subprocess(cmd=cjpeg_cmd).returncode != 0:
                 output_file.with_suffix(".jpg_").unlink(missing_ok=True)
                 return None
             output_file.with_suffix(".tga").unlink(missing_ok=True)
@@ -191,7 +194,7 @@ class SvgMinifier(BaseMinifier):
             return False
         return True
 
-    def minifier(self, cached_file: CachedFile) -> Optional[CachedFile]:
+    def minifier(self, cached_file: CachedFile) -> CachedFile | None:
         try:
             input_file = self._mkdocs_config.site_dir / cached_file.original_file_path
             output_file = self._plugin_config.cache_dir / cached_file.cached_file_name
@@ -228,7 +231,7 @@ class HtmlMinifier(BaseMinifier):
             return False
         return True
 
-    def minifier(self, cached_file: CachedFile) -> Optional[CachedFile]:
+    def minifier(self, cached_file: CachedFile) -> CachedFile | None:
         try:
             input_file = self._mkdocs_config.site_dir / cached_file.original_file_path
             output_file = self._plugin_config.cache_dir / cached_file.cached_file_name
@@ -275,7 +278,7 @@ class CssMinifier(BaseMinifier):
             return False
         return True
 
-    def minifier(self, cached_file: CachedFile) -> Optional[CachedFile]:
+    def minifier(self, cached_file: CachedFile) -> CachedFile | None:
         try:
             input_file = self._mkdocs_config.site_dir / cached_file.original_file_path
             output_file = self._plugin_config.cache_dir / cached_file.cached_file_name
@@ -315,7 +318,7 @@ class JsMinifier(BaseMinifier):
             return False
         return True
 
-    def minifier(self, cached_file: CachedFile) -> Optional[CachedFile]:
+    def minifier(self, cached_file: CachedFile) -> CachedFile | None:
         try:
             input_file = self._mkdocs_config.site_dir / cached_file.original_file_path
             output_file = self._plugin_config.cache_dir / cached_file.cached_file_name
