@@ -22,7 +22,6 @@
 
 import logging
 import re
-from typing import Optional
 
 from mkdocs_publisher.obsidian.config import _ObsidianCalloutsConfig
 
@@ -65,7 +64,7 @@ ALIGN_MAPPING = {"left": " inline", "right": " inline end"}
 class CalloutToAdmonition:
     def __init__(self, callouts_config: _ObsidianCalloutsConfig):
         self._callouts_config: _ObsidianCalloutsConfig = callouts_config
-        self._current_file_path: Optional[str] = None
+        self._current_file_path: str | None = None
 
     @staticmethod
     def _callout_indentation(match: re.Match, match_group: int, text_indentation: str = "spaces") -> str:
@@ -91,7 +90,7 @@ class CalloutToAdmonition:
             if align is None:
                 log.warning(
                     f'Wrong alignment type "{align}" in file: {self._current_file_path} '
-                    f"(fallback to no alignment, possible values {ALIGN_MAPPING.keys()})"
+                    f"(fallback to no alignment, possible values {ALIGN_MAPPING.keys()})",
                 )
                 align = ""
         else:
@@ -101,7 +100,7 @@ class CalloutToAdmonition:
         if admonition_type is None:
             log.warning(
                 f'There is no callout mapping for "{match.group(3)}" '
-                f'in file: {self._current_file_path} (fallback to "note")'
+                f'in file: {self._current_file_path} (fallback to "note")',
             )
             admonition_type = "note"
 
@@ -109,7 +108,7 @@ class CalloutToAdmonition:
         if foldable is None:
             log.warning(
                 f'Wrong definition of foldable "{match.group(4)}" '
-                f"in file: {self._current_file_path} (fallback to non foldable)"
+                f"in file: {self._current_file_path} (fallback to non foldable)",
             )
             foldable = "!!!"
 
@@ -129,14 +128,14 @@ class CalloutToAdmonition:
             if callout_match:
                 in_callout_block = True
                 markdown_lines.append(
-                    self._callout_block(match=callout_match, text_indentation=self._callouts_config.indentation)
+                    self._callout_block(match=callout_match, text_indentation=self._callouts_config.indentation),
                 )
             elif in_callout_block and callout_follow_match:
                 markdown_lines.append(
                     self._callout_block_follow(
                         match=callout_follow_match,
                         text_indentation=self._callouts_config.indentation,
-                    )
+                    ),
                 )
             else:
                 in_callout_block = False
