@@ -23,7 +23,6 @@
 import json
 import logging
 import re
-from typing import Optional
 
 import jinja2
 
@@ -62,7 +61,7 @@ class VegaCharts:
     def generate_charts(self, markdown: str) -> str:
         in_vega_block: bool = False
         vega_block_lines: list = []
-        vega_schema: Optional[str] = None
+        vega_schema: str | None = None
         markdown_lines = []
 
         for line in markdown.split("\n"):
@@ -84,14 +83,15 @@ class VegaCharts:
                         "vega_chart_id": self._vega_chart_id,
                         "vega_chart": json.dumps(vega_chart_json),
                     }
-                    vega_chart_template = jinja2.Environment(loader=jinja2.BaseLoader()).from_string(
-                        VEGA_CHART_TEMPLATE
+                    # TODO: move jinja renderrer to shared tools
+                    vega_chart_template = jinja2.Environment(loader=jinja2.BaseLoader(), autoescape=True).from_string(
+                        VEGA_CHART_TEMPLATE,
                     )
 
                     vega_chart = vega_chart_template.render(vega_chart_context)
 
                     for vega_chart_line in vega_chart.split("\n"):
-                        markdown_lines.append(vega_chart_line)
+                        markdown_lines.append(vega_chart_line)  # noqa: PERF402
 
                     # Mark that vega is in that page
                     self._vega_found = True
