@@ -22,36 +22,83 @@
 
 import logging
 from pathlib import Path
-from typing import Optional
 
 import pytest
 from pytest import LogCaptureFixture
 from pytest_check import check_functions as check
 
 from mkdocs_publisher._shared import links
+from mkdocs_publisher.meta.config import SlugModeChoiceEnum
+
+
+@pytest.mark.parametrize(
+    "text,expected",
+    {
+        ("lorem ipsum dolor", "lorem-ipsum-dolor"),
+        ("lorem_ipsum_dolor", "lorem_ipsum_dolor"),
+        ("ąćęłńóśżź", "acenoszz"),
+        ("Lorem ipsum, dolor", "lorem-ipsum-dolor"),
+    },
+)
+def test_slugify(text, expected):
+    check.equal(expected, links.slugify(text))
+
+
+@pytest.mark.parametrize(
+    "slug,title,slug_mode,warn_on_missing,expected,exp_log_level",
+    {
+        (
+            "meta_slug",
+            "title_slug",
+            SlugModeChoiceEnum.TITLE.name,
+            False,
+            "meta_slug",
+            logging.DEBUG,
+        ),
+        (None, "title_slug", SlugModeChoiceEnum.TITLE.name, False, "title_slug", logging.DEBUG),
+        (
+            "meta_slug",
+            "title_slug",
+            SlugModeChoiceEnum.FILENAME.name,
+            False,
+            "meta_slug",
+            logging.DEBUG,
+        ),
+        (None, None, SlugModeChoiceEnum.FILENAME.name, False, "file_name_slug", logging.DEBUG),
+        (None, None, SlugModeChoiceEnum.TITLE.name, True, "file_name_slug", logging.WARNING),
+        (None, None, SlugModeChoiceEnum.TITLE.name, False, "file_name_slug", logging.DEBUG),
+        (None, None, None, False, "file_name_slug", logging.DEBUG),
+    },
+)
+def test_slug_create(
+    caplog: LogCaptureFixture,
+    slug: str,
+    title: str,
+    slug_mode: str,
+    warn_on_missing: bool,
+    expected: str,
+    exp_log_level: int,
+):
+    check.equal(
+        expected,
+        links.create_slug(
+            file_name="file_name_slug",
+            slug_mode=slug_mode,
+            slug=slug,
+            title=title,
+            warn_on_missing=warn_on_missing,
+        ),
+    )
+    check.equal(exp_log_level, caplog.records[0].levelno)
 
 
 @pytest.mark.parametrize(
     "link,text,anchor,title,is_wiki,expected",
     {
         ("../file.md", "Link text", "", "", False, "[Link text](../file.md)"),
-        (
-            "../file with space.md",
-            "Link text",
-            "",
-            "",
-            False,
-            "[Link text](../file with space.md)",
-        ),
+        ("../file with space.md", "Link text", "", "", False, "[Link text](../file with space.md)"),
         ("../file", "Link text", "", "", True, "[Link text](../file.md)"),
-        (
-            "../file.md",
-            "Link text",
-            "anchor-value",
-            "",
-            False,
-            "[Link text](../file.md#anchor-value)",
-        ),
+        ("../file.md", "Link text", "anchor-value", "", False, "[Link text](../file.md#anchor-value)"),
         ("../file", "Link text", "anchor-value", "", True, "[Link text](../file.md#anchor-value)"),
         (
             "../file.md",
@@ -69,30 +116,9 @@ from mkdocs_publisher._shared import links
             True,
             '[Link text](../file.md#anchor-value "title value")',
         ),
-        (
-            "../file.md",
-            "Link text",
-            "",
-            "title value",
-            False,
-            '[Link text](../file.md "title value")',
-        ),
-        (
-            "../file",
-            "Link text",
-            "",
-            "title value",
-            True,
-            '[Link text](../file.md "title value")',
-        ),
-        (
-            None,
-            "Anchor link",
-            "just/an anchor",
-            "",
-            False,
-            "[Anchor link](#justan-anchor)",
-        ),
+        ("../file.md", "Link text", "", "title value", False, '[Link text](../file.md "title value")'),
+        ("../file", "Link text", "", "title value", True, '[Link text](../file.md "title value")'),
+        (None, "Anchor link", "just/an anchor", "", False, "[Anchor link](#justan-anchor)"),
     },
 )
 def test_link_match_dataclass(link: str, text: str, anchor: str, title: str, is_wiki: bool, expected: str):
@@ -110,7 +136,7 @@ def test_link_match_dataclass(link: str, text: str, anchor: str, title: str, is_
             "",
             "",
             "#84d6d9cdfc51cbf2e88592d12c53d5a4",
-            "[Link text](../file.md){#84d6d9cdfc51cbf2e88592d12c53d5a4}",
+            "[Link text](../file.md){ #84d6d9cdfc51cbf2e88592d12c53d5a4 }",
         ),
         (
             "../file.md",
@@ -119,7 +145,7 @@ def test_link_match_dataclass(link: str, text: str, anchor: str, title: str, is_
             "",
             "",
             "#84d6d9cdfc51cbf2e88592d12c53d5a4",
-            "[Link text](../file.md#anchor-value){#84d6d9cdfc51cbf2e88592d12c53d5a4}",
+            "[Link text](../file.md#anchor-value){ #84d6d9cdfc51cbf2e88592d12c53d5a4 }",
         ),
         (
             "../file.md",
@@ -128,7 +154,7 @@ def test_link_match_dataclass(link: str, text: str, anchor: str, title: str, is_
             "title value",
             "",
             "#84d6d9cdfc51cbf2e88592d12c53d5a4",
-            '[Link text](../file.md#anchor-value "title value")' "{#84d6d9cdfc51cbf2e88592d12c53d5a4}",
+            '[Link text](../file.md#anchor-value "title value"){ #84d6d9cdfc51cbf2e88592d12c53d5a4 }',
         ),
         (
             "../file.md",
@@ -137,34 +163,34 @@ def test_link_match_dataclass(link: str, text: str, anchor: str, title: str, is_
             "title value",
             "",
             "#84d6d9cdfc51cbf2e88592d12c53d5a4",
-            '[Link text](../file.md "title value"){#84d6d9cdfc51cbf2e88592d12c53d5a4}',
+            '[Link text](../file.md "title value"){ #84d6d9cdfc51cbf2e88592d12c53d5a4 }',
         ),
         (
             "../file.md",
             "Link text",
             "",
-            "title value",
-            "extra",
-            "#84d6d9cdfc51cbf2e88592d12c53d5a4",
-            '[Link text](../file.md "title value"){extra #84d6d9cdfc51cbf2e88592d12c53d5a4}',
-        ),
-        (
-            "../file.md",
-            "Link text",
-            "anchor-value",
-            "",
-            "extra",
-            "#84d6d9cdfc51cbf2e88592d12c53d5a4",
-            "[Link text](../file.md#anchor-value){extra #84d6d9cdfc51cbf2e88592d12c53d5a4}",
-        ),
-        (
-            "../file.md",
-            "Link text",
-            "anchor-value",
             "title value",
             "extra",
             "#84d6d9cdfc51cbf2e88592d12c53d5a4",
-            '[Link text](../file.md#anchor-value "title value")' "{extra #84d6d9cdfc51cbf2e88592d12c53d5a4}",
+            '[Link text](../file.md "title value"){ extra #84d6d9cdfc51cbf2e88592d12c53d5a4 }',
+        ),
+        (
+            "../file.md",
+            "Link text",
+            "anchor-value",
+            "",
+            "extra",
+            "#84d6d9cdfc51cbf2e88592d12c53d5a4",
+            "[Link text](../file.md#anchor-value){ extra #84d6d9cdfc51cbf2e88592d12c53d5a4 }",
+        ),
+        (
+            "../file.md",
+            "Link text",
+            "anchor-value",
+            "title value",
+            "extra",
+            "#84d6d9cdfc51cbf2e88592d12c53d5a4",
+            '[Link text](../file.md#anchor-value "title value"){ extra #84d6d9cdfc51cbf2e88592d12c53d5a4 }',
         ),
     },
 )
@@ -187,32 +213,20 @@ def test_link_match_backlinks(
     "link,image,anchor,extra,expected",
     {
         ("../image.jpg", "", "", "", "![image.jpg](../image.jpg)"),
-        (
-            "../image with space.jpg",
-            "",
-            "",
-            "",
-            "![image with space.jpg](../image with space.jpg)",
-        ),
-        ("../image.jpg", "300", "", "", "![image.jpg](../image.jpg){width=300}"),
-        ("../image.jpg", "300x300", "", "", "![image.jpg](../image.jpg){width=300 height=300}"),
-        ("../image.jpg", "", "", "align=right", "![image.jpg](../image.jpg){align=right}"),
+        ("../image with space.jpg", "", "", "", "![image with space.jpg](../image with space.jpg)"),
+        ("../image.jpg", "300", "", "", "![image.jpg](../image.jpg){ width=300 }"),
+        ("../image.jpg", "300x300", "", "", "![image.jpg](../image.jpg){ width=300 height=300 }"),
+        ("../image.jpg", "", "", "align=right", "![image.jpg](../image.jpg){ align=right }"),
         (
             "../image.jpg",
             "300x300",
             "",
             "loading=lazy",
-            "![image.jpg](../image.jpg){loading=lazy width=300 height=300}",
+            "![image.jpg](../image.jpg){ loading=lazy width=300 height=300 }",
         ),
-        ("../document.pdf", "", "", "", "![document.pdf](../document.pdf){pdfjs}"),
-        ("../document.pdf", "", "page=3", "", "![document.pdf](../document.pdf){pdfjs page=3}"),
-        (
-            "../document.pdf",
-            "",
-            "height=300",
-            "",
-            "![document.pdf](../document.pdf){pdfjs height=300}",
-        ),
+        ("../document.pdf", "", "", "", "![document.pdf](../document.pdf){ pdfjs }"),
+        ("../document.pdf", "", "page=3", "", "![document.pdf](../document.pdf){ pdfjs page=3 }"),
+        ("../document.pdf", "", "height=300", "", "![document.pdf](../document.pdf){ pdfjs height=300 }"),
     },
 )
 def test_wiki_embed_link_match_dataclass(link: str, image: str, anchor: str, extra: str, expected: str):
@@ -224,37 +238,23 @@ def test_wiki_embed_link_match_dataclass(link: str, image: str, anchor: str, ext
     "link,text,title,extra,is_loading_lazy,expected",
     {
         ("../image.jpg", "Description", "", "", False, "![Description](../image.jpg)"),
-        (
-            "../image with space.jpg",
-            "Description",
-            "",
-            "",
-            False,
-            "![Description](../image with space.jpg)",
-        ),
+        ("../image with space.jpg", "Description", "", "", False, "![Description](../image with space.jpg)"),
         (
             "../image.jpg",
             "Description",
             "",
             "align=right",
             False,
-            "![Description](../image.jpg){align=right}",
+            "![Description](../image.jpg){ align=right }",
         ),
-        (
-            "../image.jpg",
-            "Description",
-            "",
-            "",
-            True,
-            "![Description](../image.jpg){loading=lazy}",
-        ),
+        ("../image.jpg", "Description", "", "", True, "![Description](../image.jpg){ loading=lazy }"),
         (
             "../image.jpg",
             "Description",
             "",
             "align=right",
             True,
-            "![Description](../image.jpg){align=right loading=lazy}",
+            "![Description](../image.jpg){ align=right loading=lazy }",
         ),
         (
             "../image.jpg",
@@ -270,7 +270,7 @@ def test_wiki_embed_link_match_dataclass(link: str, image: str, anchor: str, ext
             "title value",
             "align=right",
             False,
-            '![Description](../image.jpg "title value"){align=right}',
+            '![Description](../image.jpg "title value"){ align=right }',
         ),
         (
             "../image.jpg",
@@ -278,12 +278,17 @@ def test_wiki_embed_link_match_dataclass(link: str, image: str, anchor: str, ext
             "title value",
             "align=right",
             True,
-            '![Description](../image.jpg "title value"){align=right loading=lazy}',
+            '![Description](../image.jpg "title value"){ align=right loading=lazy }',
         ),
     },
 )
 def test_md_embed_link_match_dataclass(
-    link: str, text: str, title: str, extra: str, is_loading_lazy: bool, expected: str
+    link: str,
+    text: str,
+    title: str,
+    extra: str,
+    is_loading_lazy: bool,
+    expected: str,
 ):
     link_obj = links.MdEmbedLinkMatch(link=link, text=text, title=title, extra=extra, is_loading_lazy=is_loading_lazy)
     assert expected == str(link_obj)
@@ -347,8 +352,8 @@ def test_relative_path_finder_multiple_file_found(
     },
 )
 def test_get_relative_file_path(
-    file_path: Optional[str],
-    expected: Optional[str],
+    file_path: str | None,
+    expected: str | None,
     test_data_dir: Path,
     relative_sub_path_finder: links.RelativePathFinder,
 ):
@@ -362,21 +367,9 @@ def test_get_relative_file_path(
     {
         ("main.md", "Main markdown", "", "", "[Main markdown](../main.md)"),
         ("main.md", "Main with anchor", "anchor", "", "[Main with anchor](../main.md#anchor)"),
-        (
-            "main.md",
-            "Main with anchor",
-            "anchor",
-            "title",
-            '[Main with anchor](../main.md#anchor "title")',
-        ),
+        ("main.md", "Main with anchor", "anchor", "title", '[Main with anchor](../main.md#anchor "title")'),
         ("#just-anchor", "Just Anchor", "", "", "[Just Anchor](#just-anchor)"),
-        (
-            "relative/other_rel_file.md",
-            "Other markdown",
-            "",
-            "",
-            "[Other markdown](other_rel_file.md)",
-        ),
+        ("relative/other_rel_file.md", "Other markdown", "", "", "[Other markdown](other_rel_file.md)"),
         ("relative/rel_sub/index-0.md", "Index 0", "", "", "[Index 0](rel_sub/index-0.md)"),
         ("relative/non_existing.md", "Non-existing", "", "", "[Non-existing]()"),
     },
@@ -397,13 +390,7 @@ def test_blog_link_match_dataclass(
 @pytest.mark.parametrize(
     "link,text,anchor,title,expected",
     {
-        (
-            "current/cur_sub/cur_sub_file.md",
-            "Sub markdown",
-            "",
-            "",
-            "[Sub markdown](current/cur_sub/cur_sub_file.md)",
-        ),
+        ("current/cur_sub/cur_sub_file.md", "Sub markdown", "", "", "[Sub markdown](current/cur_sub/cur_sub_file.md)"),
         (
             "current/cur_sub/cur_sub_file.md",
             "Sub markdown",
