@@ -22,8 +22,6 @@
 
 import logging
 from pathlib import Path
-from typing import Optional
-from typing import Union
 from unittest.mock import patch
 
 import pytest
@@ -41,8 +39,8 @@ from mkdocs_publisher.meta.plugin import MetaPlugin
 @pytest.mark.parametrize(
     "path,abs_path,expected_name,expected_parent",
     [
-        (Path("docs/fake_file.md"), Path("/Users/me/docs/fake_file.md"), "fake_file.md", Path("docs")),
-        (Path("fake_file.md"), Path("/Users/me/docs/fake_file.md"), "fake_file.md", None),
+        (Path("docs/fake_file.md"), Path("/Users/docs/fake_file.md"), "fake_file.md", Path("docs")),
+        (Path("fake_file.md"), Path("/Users/docs/fake_file.md"), "fake_file.md", None),
     ],
 )
 def test_properties(path: Path, abs_path: Path, expected_name: str, expected_parent: str):
@@ -84,9 +82,7 @@ def test_overview(
     is_dir: bool,
     expected: bool,
 ):
-    meta_file: MetaFile = MetaFile(
-        path=Path("fake_file.md"), abs_path=Path("/Users/me/docs/fake_file.md"), is_dir=is_dir
-    )
+    meta_file: MetaFile = MetaFile(path=Path("fake_file.md"), abs_path=Path("/Users/docs/fake_file.md"), is_dir=is_dir)
 
     meta_files: MetaFiles = MetaFiles()
     meta_files.set_configs(mkdocs_config=mkdocs_config, meta_plugin_config=pub_meta_plugin.config)
@@ -115,12 +111,10 @@ def test_title(
     markdown: str,
     is_dir: bool,
     expected: str,
-    warn: Optional[str],
+    warn: str | None,
     caplog: LogCaptureFixture,
 ):
-    meta_file: MetaFile = MetaFile(
-        path=Path("fake_file.md"), abs_path=Path("/Users/me/docs/fake_file.md"), is_dir=is_dir
-    )
+    meta_file: MetaFile = MetaFile(path=Path("fake_file.md"), abs_path=Path("/Users/docs/fake_file.md"), is_dir=is_dir)
 
     meta_files: MetaFiles = MetaFiles()
     meta_files.set_configs(mkdocs_config=mkdocs_config, meta_plugin_config=pub_meta_plugin.config)
@@ -149,13 +143,11 @@ def test_redirect(
     pub_meta_plugin: MetaPlugin,
     meta: dict,
     markdown: str,
-    expected_publish: Optional[str],
-    expected_redirect: Optional[str],
-    caplog: LogCaptureFixture,
+    expected_publish: str | None,
+    expected_redirect: str | None,
 ):
-    meta_file: MetaFile = MetaFile(
-        path=Path("fake_file.md"), abs_path=Path("/Users/me/docs/fake_file.md"), is_dir=False
-    )
+    meta_file: MetaFile = MetaFile(path=Path("fake_file.md"), abs_path=Path("/Users/docs/fake_file.md"), is_dir=False)
+
     meta_files: MetaFiles = MetaFiles()
     meta_files.set_configs(mkdocs_config=mkdocs_config, meta_plugin_config=pub_meta_plugin.config)
     new_meta = meta_files._get_redirect(meta_file=meta_file, meta=meta, markdown=markdown)
@@ -174,17 +166,16 @@ def test_on_serve(mkdocs_config: MkDocsConfig, pub_meta_plugin: MetaPlugin):
 
 
 @pytest.mark.parametrize(
-    "mkdocs_config,hidden_path,expected",
+    "hidden_path,expected",
     [
-        ({"docs_dir": "docs"}, None, []),
-        ({"docs_dir": "docs"}, Path("docs/hidden_dir"), [Path("hidden_dir")]),
+        (None, []),
+        (Path("docs/hidden_dir"), [Path("hidden_dir")]),
     ],
-    indirect=["mkdocs_config"],
 )
 def test_add_hidden(
     mkdocs_config: MkDocsConfig,
     pub_meta_plugin: MetaPlugin,
-    hidden_path: Optional[Path],
+    hidden_path: Path | None,
     expected: list,
 ):
     meta_files: MetaFiles = MetaFiles()
@@ -228,13 +219,13 @@ def test_get_publish_status_for_files(
     on_serve: bool,
     is_draft: bool,
     is_hidden: bool,
-    warning: Optional[str],
+    warning: str | None,
 ):
     meta_files: MetaFiles = MetaFiles()
     meta_files.on_serve = on_serve
     meta_files.set_configs(mkdocs_config=mkdocs_config, meta_plugin_config=pub_meta_plugin.config)
 
-    meta_file: MetaFile = MetaFile(path=Path("fake_file.md"), abs_path=Path("/Users/me/fake_file.md"), is_dir=False)
+    meta_file: MetaFile = MetaFile(path=Path("fake_file.md"), abs_path=Path("/Users/docs/fake_file.md"), is_dir=False)
     meta_files._get_publish_status(meta_file=meta_file, meta=meta)
 
     check.equal(meta_file.is_draft, is_draft, "Wrong draft status")
@@ -268,13 +259,13 @@ def test_get_publish_status_for_dirs(
     on_serve: bool,
     is_draft: bool,
     is_hidden: bool,
-    warning: Optional[str],
+    warning: str | None,
 ):
     meta_files: MetaFiles = MetaFiles()
     meta_files.on_serve = on_serve
     meta_files.set_configs(mkdocs_config=mkdocs_config, meta_plugin_config=pub_meta_plugin.config)
 
-    meta_file: MetaFile = MetaFile(path=Path("me"), abs_path=Path("/Users/me"), is_dir=True)
+    meta_file: MetaFile = MetaFile(path=Path("me"), abs_path=Path("/Users/docs/me"), is_dir=True)
     meta_files._get_publish_status(meta_file=meta_file, meta=meta)
 
     check.equal(meta_file.is_draft, is_draft, "Wrong draft status")
@@ -309,11 +300,13 @@ def test_get_publish_status_with_parent(
     meta_files: MetaFiles = MetaFiles()
     meta_files.set_configs(mkdocs_config=mkdocs_config, meta_plugin_config=pub_meta_plugin.config)
 
-    meta_file_parent: MetaFile = MetaFile(path=Path("me"), abs_path=Path("/Users/me"), is_dir=True)
-    meta_files["me"] = meta_file_parent
+    meta_file_parent: MetaFile = MetaFile(path=Path("docs"), abs_path=Path("/Users/docs"), is_dir=True)
+    meta_files["docs"] = meta_file_parent
     meta_files._get_publish_status(meta_file=meta_file_parent, meta=parent_meta)
 
-    meta_file: MetaFile = MetaFile(path=Path("me/fake_file.md"), abs_path=Path("/Users/me/fake_file.md"), is_dir=False)
+    meta_file: MetaFile = MetaFile(
+        path=Path("docs/fake_file.md"), abs_path=Path("/Users/docs/fake_file.md"), is_dir=False
+    )
     meta_files._get_publish_status(meta_file=meta_file, meta=meta)
 
     check.equal(meta_file.is_draft, is_draft)
@@ -321,25 +314,116 @@ def test_get_publish_status_with_parent(
 
 
 @pytest.mark.parametrize(
-    "publish_dir,publish_file,drafts_keys,draft_file_keys,draft_dir_keys,hidden_keys,hidden_file_keys,hidden_dir_keys",
+    "publish_dir,publish_file,is_overview,drafts_keys,draft_file_keys,draft_dir_keys,hidden_keys,hidden_file_keys,hidden_dir_keys",
     [
-        (False, False, {"me/fake_file.md", "me"}, {"me/fake_file.md"}, {"me"}, set(), set(), set()),
-        (False, True, {"me/fake_file.md", "me"}, {"me/fake_file.md"}, {"me"}, set(), set(), set()),
-        (False, "hidden", {"me/fake_file.md", "me"}, {"me/fake_file.md"}, {"me"}, set(), set(), set()),
-        (True, False, {"me/fake_file.md"}, {"me/fake_file.md"}, set(), set(), set(), set()),
-        (True, True, set(), set(), set(), set(), set(), set()),
-        (True, "hidden", set(), set(), set(), {"me/fake_file.md"}, {"me/fake_file.md"}, set()),
-        ("hidden", False, {"me/fake_file.md"}, {"me/fake_file.md"}, set(), {"me"}, set(), {"me"}),
-        ("hidden", True, set(), set(), set(), {"me/fake_file.md", "me"}, {"me/fake_file.md"}, {"me"}),
-        ("hidden", "hidden", set(), set(), set(), {"me/fake_file.md", "me"}, {"me/fake_file.md"}, {"me"}),
+        (False, False, False, {"docs/fake_file.md", "docs"}, {"docs/fake_file.md"}, {"docs"}, set(), set(), set()),
+        (False, True, False, {"docs/fake_file.md", "docs"}, {"docs/fake_file.md"}, {"docs"}, set(), set(), set()),
+        (False, "hidden", False, {"docs/fake_file.md", "docs"}, {"docs/fake_file.md"}, {"docs"}, set(), set(), set()),
+        (True, False, False, {"docs/fake_file.md"}, {"docs/fake_file.md"}, set(), set(), set(), set()),
+        (True, True, False, set(), set(), set(), set(), set(), set()),
+        (True, "hidden", False, set(), set(), set(), {"docs/fake_file.md"}, {"docs/fake_file.md"}, set()),
+        ("hidden", False, False, {"docs/fake_file.md"}, {"docs/fake_file.md"}, set(), {"docs"}, set(), {"docs"}),
+        ("hidden", True, False, set(), set(), set(), {"docs/fake_file.md", "docs"}, {"docs/fake_file.md"}, {"docs"}),
+        (
+            "hidden",
+            "hidden",
+            False,
+            set(),
+            set(),
+            set(),
+            {"docs/fake_file.md", "docs"},
+            {"docs/fake_file.md"},
+            {"docs"},
+        ),
+        (
+            False,
+            False,
+            True,
+            {"docs/fake_file.md", "docs"},
+            {"docs/fake_file.md", "docs/README.md"},
+            {"docs"},
+            set(),
+            set(),
+            set(),
+        ),
+        (
+            False,
+            False,
+            True,
+            {"docs/fake_file.md", "docs"},
+            {"docs/fake_file.md", "docs/README.md"},
+            {"docs"},
+            set(),
+            set(),
+            set(),
+        ),
+        (
+            False,
+            True,
+            True,
+            {"docs/fake_file.md", "docs"},
+            {"docs/fake_file.md", "docs/README.md"},
+            {"docs"},
+            set(),
+            set(),
+            set(),
+        ),
+        (
+            False,
+            "hidden",
+            True,
+            {"docs/fake_file.md", "docs"},
+            {"docs/fake_file.md", "docs/README.md"},
+            {"docs"},
+            set(),
+            set(),
+            set(),
+        ),
+        (True, False, True, {"docs/fake_file.md"}, {"docs/fake_file.md"}, set(), set(), set(), set()),
+        (True, True, True, set(), set(), set(), set(), set(), set()),
+        (True, "hidden", True, set(), set(), set(), {"docs/fake_file.md"}, {"docs/fake_file.md"}, set()),
+        (
+            "hidden",
+            False,
+            True,
+            {"docs/fake_file.md"},
+            {"docs/fake_file.md"},
+            set(),
+            {"docs"},
+            {"docs/README.md"},
+            {"docs"},
+        ),
+        (
+            "hidden",
+            True,
+            True,
+            set(),
+            set(),
+            set(),
+            {"docs/fake_file.md", "docs"},
+            {"docs/fake_file.md", "docs/README.md"},
+            {"docs"},
+        ),
+        (
+            "hidden",
+            "hidden",
+            True,
+            set(),
+            set(),
+            set(),
+            {"docs/fake_file.md", "docs"},
+            {"docs/fake_file.md", "docs/README.md"},
+            {"docs"},
+        ),
     ],
 )
 def test_drafts_and_hidden(
     mkdocs_config: MkDocsConfig,
     pub_meta_plugin: MetaPlugin,
+    is_overview: bool,
     patched_meta_files: MetaFiles,
-    publish_dir: Union[str, bool],
-    publish_file: Union[str, bool],
+    publish_dir: str | bool,
+    publish_file: str | bool,
     drafts_keys: list[str],
     draft_file_keys: list[str],
     draft_dir_keys: list[str],
@@ -349,12 +433,16 @@ def test_drafts_and_hidden(
 ):
     patched_meta_files.set_configs(mkdocs_config=mkdocs_config, meta_plugin_config=pub_meta_plugin.config)
 
-    meta_file_dir: MetaFile = MetaFile(path=Path("me"), abs_path=Path("/Users/me"), is_dir=True)
-    patched_meta_files["me"] = meta_file_dir
+    meta_file_dir: MetaFile = MetaFile(
+        path=Path("docs"), abs_path=Path("/Users/docs"), is_dir=True, is_overview=is_overview
+    )
+    patched_meta_files["docs"] = meta_file_dir
     patched_meta_files._get_publish_status(meta_file=meta_file_dir, meta={"publish": publish_dir})
 
-    meta_file: MetaFile = MetaFile(path=Path("me/fake_file.md"), abs_path=Path("/Users/me/fake_file.md"), is_dir=False)
-    patched_meta_files["me/fake_file.md"] = meta_file
+    meta_file: MetaFile = MetaFile(
+        path=Path("docs/fake_file.md"), abs_path=Path("/Users/docs/fake_file.md"), is_dir=False
+    )
+    patched_meta_files["docs/fake_file.md"] = meta_file
     patched_meta_files._get_publish_status(meta_file=meta_file, meta={"publish": publish_file})
 
     check.equal(set(patched_meta_files.drafts.keys()), drafts_keys, "Wrong drafts keys")
@@ -365,25 +453,60 @@ def test_drafts_and_hidden(
     check.equal(set(patched_meta_files.hidden_dirs.keys()), hidden_dir_keys, "Wrong hidden dirs keys")
 
 
+@pytest.mark.parametrize("is_draft,expected_files", [(False, 1), (True, 0)])
+def test_clean_draft_files(
+    mkdocs_config: MkDocsConfig,
+    pub_meta_plugin: MetaPlugin,
+    patched_meta_files: MetaFiles,
+    caplog,
+    is_draft: bool,
+    expected_files: int,
+):
+    patched_meta_files.set_configs(mkdocs_config=mkdocs_config, meta_plugin_config=pub_meta_plugin.config)
+    meta_file: MetaFile = MetaFile(path=Path("fake_file.md"), abs_path=Path("/docs/fake_file.md"), is_dir=False)
+    patched_meta_files["fake_file.md"] = meta_file
+    patched_meta_files["fake_file.md"].is_draft = is_draft
+
+    files = Files(
+        files=[
+            File(
+                path="fake_file.md",
+                src_dir="docs/",
+                dest_dir="site/",
+                use_directory_urls=True,
+            )
+        ]
+    )
+
+    new_files = patched_meta_files.clean_draft_files(files=files)
+
+    check.equal(len(new_files), expected_files, "Wrong number of files")
+    if expected_files == 0:
+        check.equal(caplog.records[-1].levelno, logging.DEBUG, "Wrong log level")
+        check.is_true(caplog.records[-1].message.startswith("Removed draft file:"), "Wrong log message")
+    elif expected_files == 1:
+        check.equal(next(iter(new_files.src_paths.keys())), "fake_file.md")
+
+
 @pytest.mark.parametrize("exists", [True, False])
 def test_add_dir(mkdocs_config: MkDocsConfig, pub_meta_plugin: MetaPlugin, exists, patched_meta_files: MetaFiles):
     patched_meta_files.set_configs(mkdocs_config=mkdocs_config, meta_plugin_config=pub_meta_plugin.config)
 
-    meta_file_dir: MetaFile = MetaFile(path=Path("me"), abs_path=Path("/Users/me"), is_dir=True)
+    meta_file_dir: MetaFile = MetaFile(path=Path("docs"), abs_path=Path("/Users/docs"), is_dir=True)
     with patch.object(Path, "exists", return_value=exists):
-        patched_meta_files["me"] = meta_file_dir
+        patched_meta_files["docs"] = meta_file_dir
         patched_meta_files._get_publish_status(meta_file=meta_file_dir, meta={})
 
-    check.is_true("me" in patched_meta_files)
+    check.is_true("docs" in patched_meta_files)
 
 
 @pytest.mark.parametrize(
     "path,is_dir,ignored_dirs,expected",
     [
-        ("/Users/me/fake_file.md", False, [], ["fake_file.md"]),
-        ("/Users/me/fake_dir", True, [], ["fake_dir"]),
-        ("/Users/me/fake_dir", True, ["/Users/me/fake_dir"], []),
-        ("/Users/me/no_md.pic", False, ["fake_dir"], []),
+        ("docs/fake_file.md", False, [], ["fake_file.md"]),
+        ("docs/fake_dir", True, [], ["fake_dir"]),
+        ("docs/fake_dir", True, ["docs/fake_dir"], []),
+        ("docs/no_md.pic", False, ["fake_dir"], []),
     ],
 )
 def test_add_meta_files(
@@ -400,7 +523,7 @@ def test_add_meta_files(
         patch.object(Path, "is_dir", return_value=is_dir),
         patch.object(Path, "rglob", return_value=[Path(path)]),
     ):
-        patched_meta_files.add_meta_files(ignored_dirs=[Path(ignored_dir) for ignored_dir in ignored_dirs])
+        patched_meta_files.add_files(ignored_dirs=[Path(ignored_dir) for ignored_dir in ignored_dirs])
 
     check.equal(list(patched_meta_files.keys()), expected)
 
@@ -411,17 +534,15 @@ def test_files_gen(
     patched_meta_files: MetaFiles,
 ):
     patched_meta_files.set_configs(mkdocs_config=mkdocs_config, meta_plugin_config=pub_meta_plugin.config)
-    files_paths: list[str] = ["/Users/me/fake_data.md", "/Users/me/fake_file.md"]
+    files_paths: list[str] = ["docs/fake_data.md", "docs/fake_file.md"]
     for path in files_paths:
         with (
             patch.object(Path, "is_dir", return_value=False),
             patch.object(Path, "rglob", return_value=[Path(path)]),
         ):
-            patched_meta_files.add_meta_files(ignored_dirs=[])
+            patched_meta_files.add_files(ignored_dirs=[])
 
-    gen_files_paths = []
-    for file in patched_meta_files.files_gen():
-        gen_files_paths.append(str(file.abs_path))
+    gen_files_paths = [str(file.abs_path) for file in patched_meta_files.files_gen()]
 
     check.equal(files_paths, gen_files_paths)
 
@@ -438,13 +559,13 @@ def test_clean_redirect_files(
     mkdocs_config: MkDocsConfig,
     pub_meta_plugin: MetaPlugin,
     patched_meta_files: MetaFiles,
-    redirect: Optional[str],
+    redirect: str | None,
     expected_nr_of_files: int,
 ):
     patched_meta_files.set_configs(mkdocs_config=mkdocs_config, meta_plugin_config=pub_meta_plugin.config)
     meta_file: MetaFile = MetaFile(
         path=Path("fake_url_redirect.md"),
-        abs_path=Path("/Users/me/fake_url_redirect.md"),
+        abs_path=Path("docs/fake_url_redirect.md"),
         is_dir=False,
     )
 
@@ -455,8 +576,8 @@ def test_clean_redirect_files(
         files=[
             File(
                 path="fake_url_redirect.md",
-                src_dir="/Users/me/",
-                dest_dir="/Users/docs",
+                src_dir="docs/",
+                dest_dir="site",
                 use_directory_urls=True,
             )
         ]
@@ -479,14 +600,14 @@ def test_generate_redirect_page(
     mkdocs_config: MkDocsConfig,
     pub_meta_plugin: MetaPlugin,
     patched_meta_files: MetaFiles,
-    redirect: Optional[str],
-    expected_result: Optional[str],
+    redirect: str | None,
+    expected_result: str | None,
 ):
     patched_meta_files.set_configs(mkdocs_config=mkdocs_config, meta_plugin_config=pub_meta_plugin.config)
 
     meta_file: MetaFile = MetaFile(
         path=Path("fake_url_redirect.md"),
-        abs_path=Path("/Users/me/fake_url_redirect.md"),
+        abs_path=Path("docs/fake_url_redirect.md"),
         is_dir=False,
     )
 
@@ -496,8 +617,8 @@ def test_generate_redirect_page(
 
     file = File(
         path="fake_url_redirect.md",
-        src_dir="/Users/me/",
-        dest_dir="/Users/docs",
+        src_dir="docs/",
+        dest_dir="site/",
         use_directory_urls=True,
     )
 
@@ -515,14 +636,14 @@ def test_generate_redirect_page(
         ("fake_file.md", True, "file-slug", None, "file-slug/"),
         ("fake_file.md", False, None, None, "fake_file/"),
         ("fake_file.jpeg", False, None, None, "fake_file.jpeg"),
-        ("me/fake_file.md", True, None, None, "me/fake-file/"),
-        ("me/fake_file.md", True, "file-slug", None, "me/file-slug/"),
-        ("me/fake_file.md", False, None, None, "me/fake_file/"),
-        ("me/fake_file.jpeg", False, None, None, "me/fake_file.jpeg"),
-        ("me/fake_file.md", True, None, "dir-slug", "dir-slug/fake-file/"),
-        ("me/fake_file.md", True, "file-slug", "dir-slug", "dir-slug/file-slug/"),
-        ("me/fake_file.md", False, None, "dir-slug", "dir-slug/fake_file/"),
-        ("me/fake_file.jpeg", False, None, "dir-slug", "dir-slug/fake_file.jpeg"),
+        ("docs/fake_file.md", True, None, None, "docs/fake-file/"),
+        ("docs/fake_file.md", True, "file-slug", None, "docs/file-slug/"),
+        ("docs/fake_file.md", False, None, None, "docs/fake_file/"),
+        ("docs/fake_file.jpeg", False, None, None, "docs/fake_file.jpeg"),
+        ("docs/fake_file.md", True, None, "dir-slug", "dir-slug/fake-file/"),
+        ("docs/fake_file.md", True, "file-slug", "dir-slug", "dir-slug/file-slug/"),
+        ("docs/fake_file.md", False, None, "dir-slug", "dir-slug/fake_file/"),
+        ("docs/fake_file.jpeg", False, None, "dir-slug", "dir-slug/fake_file.jpeg"),
         (".", False, None, None, "."),
     ],
 )
@@ -532,26 +653,26 @@ def test_change_file_slug(
     patched_meta_files: MetaFiles,
     path: str,
     with_meta_file: bool,
-    file_slug: Optional[str],
-    dir_slug: Optional[str],
-    expected_url: Optional[str],
+    file_slug: str | None,
+    dir_slug: str | None,
+    expected_url: str | None,
 ):
     patched_meta_files.set_configs(mkdocs_config=mkdocs_config, meta_plugin_config=pub_meta_plugin.config)
 
     meta_dir: MetaFile = MetaFile(
-        path=Path("me"),
-        abs_path=Path("/Users/me/README.md"),
+        path=Path("docs"),
+        abs_path=Path("/Users/docs/README.md"),
         is_dir=True,
     )
     with patch.object(Path, "exists", return_value=True):
-        patched_meta_files["me"] = meta_dir
+        patched_meta_files["docs"] = meta_dir
     if dir_slug:
         meta_dir.slug = dir_slug
 
     if with_meta_file:
         meta_file: MetaFile = MetaFile(
             path=Path(path),
-            abs_path=Path(f"/Users/me/{'index.md' if path == '.' else path}"),
+            abs_path=Path(f"/Users/docs/{'index.md' if path == '.' else path}"),
             is_dir=False,
         )
         with patch.object(Path, "exists", return_value=True):
@@ -561,8 +682,8 @@ def test_change_file_slug(
 
     file = File(
         path=path,
-        src_dir="/Users",
-        dest_dir="/Users/out",
+        src_dir="/Users/docs",
+        dest_dir="/Users/site",
         use_directory_urls=True,
     )
 
@@ -577,7 +698,7 @@ def test_change_file_slug(
         ({}, "fake_file.md", True, False, "/Users/me", "fake-file/"),
         ({}, "fake_file.md", False, False, None, "fake_file/"),
         ({}, "fake_file.md", True, True, None, None),
-        ({}, "me/fake_file.md", True, False, "/Users/me", None),
+        ({}, "docs/fake_file.md", True, False, "/Users/docs", None),
         ({"slug": {"enabled": False}}, "fake_file.md", True, False, None, "fake_file/"),
     ],
     indirect=["pub_meta_plugin"],
@@ -589,23 +710,24 @@ def test_change_files_slug(
     path: str,
     with_meta_file: bool,
     is_draft: bool,
-    ignored_dir: Optional[str],
-    expected_url: Optional[str],
+    ignored_dir: str | None,
+    expected_url: str | None,
 ):
+    patched_meta_files = MetaFiles()
     patched_meta_files.set_configs(mkdocs_config=mkdocs_config, meta_plugin_config=pub_meta_plugin.config)
 
     meta_dir: MetaFile = MetaFile(
-        path=Path("me"),
-        abs_path=Path("/Users/me/README.md"),
+        path=Path("docs"),
+        abs_path=Path("/Users/docs/README.md"),
         is_dir=True,
     )
     with patch.object(Path, "exists", return_value=True):
-        patched_meta_files["me"] = meta_dir
+        patched_meta_files["docs"] = meta_dir
 
     if with_meta_file:
         meta_file: MetaFile = MetaFile(
             path=Path(path),
-            abs_path=Path(f"/Users/me/{'index.md' if path == '.' else path}"),
+            abs_path=Path(f"/Users/docs/{'index.md' if path == '.' else path}"),
             is_dir=False,
         )
         with patch.object(Path, "exists", return_value=True):
@@ -616,8 +738,8 @@ def test_change_files_slug(
         [
             File(
                 path=path,
-                src_dir="/Users",
-                dest_dir="/Users/out",
+                src_dir="/Users/docs",
+                dest_dir="/Users/site",
                 use_directory_urls=True,
             )
         ]
