@@ -37,19 +37,19 @@ class HTMLModifier:
 
     def add_head_script(self, src: str):
         """Add script to the head section"""
-        self._soup.head.append(  # type: ignore
-            self._soup.new_tag(name="script", attrs={"src": src})
+        self._soup.head.append(
+            self._soup.new_tag(name="script", attrs={"src": src}),
         )
 
     def add_meta_property(self, name: str, value: str):
         """Add property to the head section"""
-        self._soup.head.append(  # type: ignore
-            self._soup.new_tag(name="meta", attrs={"property": name, "content": value})
+        self._soup.head.append(
+            self._soup.new_tag(name="meta", attrs={"property": name, "content": value}),
         )
 
     def remove_meta_properties(self, properties: list[str]):
         """Remove meta tags with given properties from head section"""
         for prop in properties:
-            head_property = self._soup.head.find(name="meta", attr={"property": prop})  # type: ignore
+            head_property = self._soup.head.find(name="meta", attr={"property": prop})
             if head_property is not None:
                 head_property.extract()
