@@ -25,22 +25,22 @@ import platform
 import sys
 from io import BytesIO
 from pathlib import Path
-from typing import cast
 from zipfile import ZIP_DEFLATED
 from zipfile import ZipFile
 
+from mkdocs.config.defaults import MkDocsConfig
 from mkdocs.plugins import BasePlugin
 from mkdocs.plugins import event_priority
 
 from mkdocs_publisher._shared import file_utils
 from mkdocs_publisher._shared import mkdocs_utils
 from mkdocs_publisher.debugger import loggers
-from mkdocs_publisher.debugger.config import DebuggerConfig
+from mkdocs_publisher.debugger.config import DebuggerPluginConfig
 
 log = logging.getLogger("mkdocs.publisher.debug.plugin")
 
-LOG_FILENAME_SUFFIX = "_mkdocs_build.log"
-ZIP_FILENAME_SUFFIX = "_mkdocs_debug.zip"
+LOG_FILENAME_SUFFIX = "_mkdocs_build.log"  # TODO: make it configurable
+ZIP_FILENAME_SUFFIX = "_mkdocs_debug.zip"  # TODO: make it configurable
 FILES_TO_ZIP_LIST = [
     "mkdocs.yml",
     "requirements.txt",
@@ -48,64 +48,71 @@ FILES_TO_ZIP_LIST = [
     "poetry.lock",
     ".gitignore",
 ]
-PIP_FREEZE_FILENAME = "requirements_freeze.txt"
+PIP_FREEZE_FILENAME = "requirements_pub_debugger.txt"  # TODO: make it configurable
 
 
-class DebuggerPlugin(BasePlugin[DebuggerConfig]):
+class DebuggerPlugin(BasePlugin[DebuggerPluginConfig]):
+    supports_multiple_instances = False
+
     def __init__(self):
         self._mkdocs_log_stream_handler: logging.Handler = logging.getLogger("mkdocs").handlers[0]
 
         self._mkdocs_log_file_handler: loggers.DatedFileHandler = loggers.DatedFileHandler(
-            filename=f"%Y%m%d_%H%M%S{LOG_FILENAME_SUFFIX}"
+            filename=f"%Y%m%d_%H%M%S{LOG_FILENAME_SUFFIX}",
         )
 
         self._mkdocs_log_file: str = str(Path(self._mkdocs_log_file_handler.baseFilename).name)
         self._mkdocs_log_date: str = self._mkdocs_log_file.replace(LOG_FILENAME_SUFFIX, "")
+        self._mkdocs_config: MkDocsConfig = mkdocs_utils.get_mkdocs_config()
 
-        self.load_config(
-            options=cast(
-                dict,
-                mkdocs_utils.get_plugin_config(
-                    mkdocs_config=mkdocs_utils.get_mkdocs_config(), plugin_name="pub-debugger"
-                ),
-            ),
-            config_file_path=mkdocs_utils.get_mkdocs_config().config_file_path,
-        )
+        self.config: DebuggerPluginConfig = mkdocs_utils.get_plugin_config(
+            plugin_config_type=DebuggerPluginConfig,  # type: ignore[reportArgumentType]
+            mkdocs_config=self._mkdocs_config,
+        )  # type: ignore[reportAttributeAccessIssue]
 
         if self.config.console_log.enabled:
             self._mkdocs_log_stream_handler.setFormatter(
-                loggers.ProjectPathStreamFormatter(console_config=self.config.console_log)
+                loggers.ProjectPathStreamFormatter(console_config=self.config.console_log),
             )
             self._mkdocs_log_stream_handler.addFilter(
-                loggers.ProjectPathConsoleFilter(console_config=self.config.console_log)
+                loggers.ProjectPathConsoleFilter(console_config=self.config.console_log),
             )
             # noinspection PyUnresolvedReferences
-            self._mkdocs_log_stream_handler.level = logging._nameToLevel[self.config.console_log.log_level]
+            mkdocs_log_level = logging.getLogger("mkdocs").level
+            console_log_level = logging._nameToLevel[self.config.console_log.log_level]  # noqa: SLF001
 
+            if mkdocs_log_level <= console_log_level:
+                console_log_level = mkdocs_log_level
+
+            self._mkdocs_log_stream_handler.setLevel(console_log_level)
             logging.getLogger("root").handlers = [self._mkdocs_log_stream_handler]
+            logging.getLogger("mkdocs").setLevel(console_log_level)
 
         if self.config.file_log.enabled:
             self._mkdocs_log_file_handler.setFormatter(
-                loggers.ProjectPathFileFormatter(fmt=self.config.file_log.log_format)
+                loggers.ProjectPathFileFormatter(
+                    fmt=self.config.file_log.log_format,
+                    datefmt=self.config.file_log.time_format,
+                ),
             )
             self._mkdocs_log_file_handler.addFilter(loggers.ProjectPathFileFilter(file_config=self.config.file_log))
             # noinspection PyUnresolvedReferences
-            self._mkdocs_log_file_handler.level = logging._nameToLevel[self.config.file_log.log_level]
+            self._mkdocs_log_file_handler.setLevel(logging._nameToLevel[self.config.file_log.log_level])  # noqa: SLF001
 
             logging.getLogger("mkdocs").handlers.append(self._mkdocs_log_file_handler)
 
             if self.config.file_log.remove_old_files:
-                for log_file in Path(".").rglob(f"*{LOG_FILENAME_SUFFIX}"):
+                for log_file in Path().rglob(f"*{LOG_FILENAME_SUFFIX}"):
                     if str(log_file) != self._mkdocs_log_file:
                         log_file.unlink(missing_ok=True)
 
-    @event_priority(100)  # Run after  all other plugins
+    @event_priority(100)  # Run after all other plugins
     def on_shutdown(self) -> None:
         if self.config.file_log.enabled:
             log.info(f"Platform: {platform.platform()}")
             log.info(
                 f"Python version: {platform.python_version()} "
-                f"(using virtual environment: {str(sys.prefix != sys.base_prefix).lower()})"
+                f"(using virtual environment: {str(sys.prefix != sys.base_prefix).lower()})",
             )
             log.info(f"Build log file: {self._mkdocs_log_file_handler.baseFilename}")
 
@@ -113,7 +120,7 @@ class DebuggerPlugin(BasePlugin[DebuggerConfig]):
             zip_file_name = f"{self._mkdocs_log_date}{ZIP_FILENAME_SUFFIX}"
 
             if self.config.zip_log.remove_old_files:
-                for log_file in Path(".").rglob(f"*{ZIP_FILENAME_SUFFIX}"):
+                for log_file in Path().rglob(f"*{ZIP_FILENAME_SUFFIX}"):
                     if str(log_file) != zip_file_name:
                         log_file.unlink(missing_ok=True)
 
@@ -138,6 +145,6 @@ class DebuggerPlugin(BasePlugin[DebuggerConfig]):
                 # Write build log file
                 archive_file.write(filename=self._mkdocs_log_file, arcname=self._mkdocs_log_file)
 
-            with open(zip_file_name, "wb") as zip_file:
+            with Path(zip_file_name).open("wb") as zip_file:
                 zip_file.write(archive.getvalue())
                 log.info(f"Debugger ZIP file: {zip_file_name}")
